Move Site model to wagtail.core.models.sites
